Update README.md
I was getting frames after every 15 frames prior to this edit. But it was producing a lot of redundancy.
Further, I was zooming 0.2 prior to this. But in my video some of the frames were zoomed by doctor for better explanation so there was a chance that zooming by 0.2 maybe cropped out the ROI.